ch06_rag_handson/requirements.txt
boto3
langchain==0.3.21
langchain-aws==0.2.18
ragas==0.2.14
datasets==3.0.2
langchain-text-splitters
mypy-boto3-bedrock-runtime
mypy-boto3-bedrock-agent
mypy-boto3-bedrock-agent-runtime
streamlit==1.39.0
python-dotenv==1.0.1